newsletter/requirements.txt
pika>=1.3.0,<1.4.0
python-dotenv==0.20.0
